Revert "Merge upstream 10-25"
