+++ lib/phobos/actions/process_batch.rb
@@ -4,7 +4,6 @@ module Phobos
   module Actions
     class ProcessBatch
       include Phobos::Instrumentation
-      include Phobos::Log
 
       attr_reader :metadata
 
@@ -27,11 +26,7 @@ module Phobos
               message: message,
               listener_metadata: @listener_metadata
             ).execute
-            begin
-              @listener.consumer.trigger_heartbeat
-            rescue Kafka::HeartbeatError => e
-              log_warn("Error sending Heartbeat #{e.class.name}-#{e}")
-            end
+            @listener.consumer.trigger_heartbeat
           end
         end
       end
