+++ lib/phobos/handler.rb
@@ -2,7 +2,6 @@
 
 module Phobos
   module Handler
-    # @!visibility private
     def self.included(base)
       base.extend(ClassMethods)
     end
