+++ lib/phobos/listener.rb
@@ -9,28 +9,9 @@ module Phobos
     DEFAULT_MAX_BYTES_PER_PARTITION = 1_048_576 # 1 MB
     DELIVERY_OPTS = %w[batch message inline_batch].freeze
 
-    # @return [String]
-    attr_reader :group_id
-    # @return [String]
-    attr_reader :topic
-    attr_reader :id
-    # @return [Class]
-    attr_reader :handler_class
-    attr_reader :encoding, :consumer
+    attr_reader :group_id, :topic, :id
+    attr_reader :handler_class, :encoding, :consumer
 
-    # @param handler [Class]
-    # @param group_id [String]
-    # @param topic [String]
-    # @param min_bytes [Integer]
-    # @param max_wait_time [Integer]
-    # @param start_from_beginning [Boolean]
-    # @param delivery [String]
-    # @param max_bytes_per_partition [Integer]
-    # @param session_timeout [Integer]
-    # @param offset_commit_interval [Integer]
-    # @param heartbeat_interval [Integer]
-    # @param offset_commit_threshold [Integer]
-    # @param offset_retention_time [Integer]
     # rubocop:disable Metrics/MethodLength
     def initialize(handler:, group_id:, topic:, min_bytes: nil, max_wait_time: nil,
                    force_encoding: nil, start_from_beginning: true, backoff: nil,
@@ -59,14 +40,14 @@ module Phobos
     end
     # rubocop:enable Metrics/MethodLength
 
-    # @return [void]
     def start
       @signal_to_stop = false
+
       start_listener
 
       begin
         start_consumer_loop
-      rescue Kafka::ProcessingError, Phobos::AbortError => e
+      rescue Kafka::ProcessingError, Phobos::AbortError
         # Abort is an exception to prevent the consumer from committing the offset.
         # Since "listener" had a message being retried while "stop" was called
         # it's wise to not commit the batch offset to avoid data loss. This will
@@ -74,13 +55,11 @@ module Phobos
         instrument('listener.retry_aborted', listener_metadata) do
           log_info('Retry loop aborted, listener is shutting down', listener_metadata)
         end
-        raise e if e.is_a?(Kafka::ProcessingError)
       end
     ensure
       stop_listener
     end
 
-    # @return [void]
     def stop
       return if should_stop?
