+++ lib/phobos/producer.rb
@@ -2,13 +2,11 @@
 
 module Phobos
   module Producer
-    # @!visibility private
     def self.included(base)
       base.extend(Phobos::Producer::ClassMethods)
       base.class_variable_set(:@@producer_config, :producer)
     end
 
-    # @return [Phobos::Producer::PublicAPI]
     def producer
       Phobos::Producer::PublicAPI.new(self)
     end
@@ -18,12 +16,6 @@ module Phobos
         @host_obj = host_obj
       end
 
-      # @param topic [String]
-      # @param payload [String]
-      # @param key [String]
-      # @param partition_key [Integer]
-      # @param headers [Hash]
-      # @return [void]
       def publish(topic:, payload:, key: nil, partition_key: nil, headers: nil)
         class_producer.publish(topic: topic,
                                payload: payload,
@@ -32,12 +24,6 @@ module Phobos
                                headers: headers)
       end
 
-      # @param topic [String]
-      # @param payload [String]
-      # @param key [String]
-      # @param partition_key [Integer]
-      # @param headers [Hash]
-      # @return [void]
       def async_publish(topic:, payload:, key: nil, partition_key: nil, headers: nil)
         class_producer.async_publish(topic: topic,
                                      payload: payload,
@@ -46,7 +32,7 @@ module Phobos
                                      headers: headers)
       end
 
-      # @param messages [Array<Hash>]
+      # @param messages [Array(Hash(:topic, :payload, :key, :headers))]
       #        e.g.: [
       #          { topic: 'A', payload: 'message-1', key: '1', headers: { foo: 'bar' } },
       #          { topic: 'B', payload: 'message-2', key: '2', headers: { foo: 'bar' } }
@@ -56,7 +42,6 @@ module Phobos
         class_producer.publish_list(messages)
       end
 
-      # @param messages [Array<Hash>]
       def async_publish_list(messages)
         class_producer.async_publish_list(messages)
       end
@@ -69,7 +54,6 @@ module Phobos
     end
 
     module ClassMethods
-      # @return [Phobos::Producer::ClassMethods::PublicAPI]
       def producer
         Phobos::Producer::ClassMethods::PublicAPI.new(self)
       end
@@ -83,11 +67,8 @@ module Phobos
       end
 
       class PublicAPI
-        # @return [Symbol]
         NAMESPACE = :phobos_producer_store
-        # @return [Array<Symbol>]
         ASYNC_PRODUCER_PARAMS = [:max_queue_size, :delivery_threshold, :delivery_interval].freeze
-        # @return [Array<Symbol>]
         INTERNAL_PRODUCER_PARAMS = [:persistent_connections].freeze
 
         def initialize(host_class)
@@ -98,18 +79,16 @@ module Phobos
         # performed by the host class
         #
         # @param kafka_client [Kafka::Client]
-        # @return [void]
+        #
         def configure_kafka_client(kafka_client)
           async_producer_shutdown
           producer_store[:kafka_client] = kafka_client
         end
 
-        # @return [Kafka::Client]
         def kafka_client
           producer_store[:kafka_client]
         end
 
-        # @return [Kafka::Producer]
         def create_sync_producer
           client = kafka_client || configure_kafka_client(create_kafka_client)
           sync_producer = client.producer(**regular_configs)
@@ -119,29 +98,20 @@ module Phobos
           sync_producer
         end
 
-        # @return [Kafka::Producer]
         def sync_producer
           producer_store[:sync_producer]
         end
 
-        # @return [void]
         def sync_producer_shutdown
           sync_producer&.shutdown
           producer_store[:sync_producer] = nil
         end
 
-        # @param topic [String]
-        # @param payload [String]
-        # @param partition_key [Integer]
-        # @param headers [Hash]
-        # @return [void]
         def publish(topic:, payload:, key: nil, partition_key: nil, headers: nil)
           publish_list([{ topic: topic, payload: payload, key: key,
                           partition_key: partition_key, headers: headers }])
         end
 
-        # @param messages [Array<Hash>]
-        # @return [void]
         def publish_list(messages)
           producer = sync_producer || create_sync_producer
           produce_messages(producer, messages)
@@ -150,51 +120,39 @@ module Phobos
           producer&.shutdown unless Phobos.config.producer_hash[:persistent_connections]
         end
 
-        # @return [Kafka::AsyncProducer]
         def create_async_producer
           client = kafka_client || configure_kafka_client(create_kafka_client)
           async_producer = client.async_producer(**async_configs)
           producer_store[:async_producer] = async_producer
         end
 
-        # @return [Kafka::AsyncProducer]
         def async_producer
           producer_store[:async_producer]
         end
 
-        # @param topic [String]
-        # @param payload [String]
-        # @param partition_key [Integer]
-        # @param headers [Hash]
-        # @return [void]
         def async_publish(topic:, payload:, key: nil, partition_key: nil, headers: nil)
           async_publish_list([{ topic: topic, payload: payload, key: key,
                                 partition_key: partition_key, headers: headers }])
         end
 
-        # @param messages [Array<Hash>]
-        # @return [void]
         def async_publish_list(messages)
           producer = async_producer || create_async_producer
           produce_messages(producer, messages)
           producer.deliver_messages unless async_automatic_delivery?
         end
 
-        # @return [void]
         def async_producer_shutdown
           async_producer&.deliver_messages
           async_producer&.shutdown
           producer_store[:async_producer] = nil
         end
 
-        # @return [Hash]
         def regular_configs
           Phobos.config.producer_hash
                 .reject { |k, _| ASYNC_PRODUCER_PARAMS.include?(k) }
                 .reject { |k, _| INTERNAL_PRODUCER_PARAMS.include?(k) }
         end
 
-        # @return [Hash]
         def async_configs
           Phobos.config.producer_hash
                 .reject { |k, _| INTERNAL_PRODUCER_PARAMS.include?(k) }
