+++ lib/phobos/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Phobos
-  VERSION = '3.1.0'
+  VERSION = '3.0.4'
 end
