--- sig/kafka.rbs
@@ -1,12 +1,0 @@
-# Ruby-Kafka definitions
-
-module Kafka
-  class Client
-  end
-
-  class Producer
-  end
-
-  class AsyncProducer
-  end
-end
